Create log.txt
Place to log any mistake or error the program made while it was running. Preferably use the errorLog(String message, boolean critic) function to write to this file, it will make a print() and write it for you with the date and a request to talk about the issue to the development team.